pyproject.toml
[tool.poetry]
name = "Python_youtube_downloader"
version = "0.1.0"
description = ""
authors = ["Francisco Javier Asensi Benito <asensi.benito@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytube3 = "^9.6.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
